Mimic move in develop_radar on master, to allow merge
