+++ stonesoup/sensor/radar/tests/test_radar.py
@@ -3,10 +3,10 @@ import datetime
 
 import numpy as np
 
-from ...functions import cart2pol
-from ...types.angle import Bearing
-from ...types.array import StateVector, CovarianceMatrix
-from ...types.state import State
+from ....functions import cart2pol
+from ....types.angle import Bearing
+from ....types.array import StateVector, CovarianceMatrix
+from ....types.state import State
 from ..radar import (
     RadarRangeBearing, RadarRotatingRangeBearing, RadarRasterScanRangeBearing)
